+++ configuration/config.js
@@ -58,6 +58,9 @@ config.ld_library_path = config.ossec_path + "/framework/lib"
 // Option to force the use of authd to remove and add agents
 config.use_only_authd = false;
 
+// Option to drop privileges (run as ossec)
+config.drop_privileges = true;
+
 /************************* SSL OPTIONS ****************************************/
 // SSL protocol