+++ controllers/agents.js
@@ -39,7 +39,11 @@ router.get('/', cache(), function(req, res) {
     req.apicacheGroup = "agents";
 
     var data_request = {'function': '/agents', 'arguments': {}};
-    var filters = {'offset': 'numbers', 'limit': 'numbers', 'sort':'sort_param', 'select':'select_param', 'search':'search_param', 'status':'alphanumeric_param', 'os.platform':'alphanumeric_param', 'os.version':'alphanumeric_param', 'manager':'alphanumeric_param', 'version':'alphanumeric_param'};
+    var filters = {'offset': 'numbers', 'limit': 'numbers', 'sort':'sort_param',
+                   'select':'select_param', 'search':'search_param',
+                   'status':'alphanumeric_param', 'os.platform':'alphanumeric_param',
+                   'os.version':'alphanumeric_param', 'manager':'alphanumeric_param',
+                   'version':'alphanumeric_param', 'node': 'alphanumeric_param'};
 
     if (!filter.check(req.query, filters, req, res))  // Filter with error
         return;
@@ -62,6 +66,8 @@ router.get('/', cache(), function(req, res) {
         data_request['arguments']['os_version'] = req.query['os.version'];
     if ('manager' in req.query)
         data_request['arguments']['manager_host'] = req.query['manager'];
+    if ('node' in req.query)
+        data_request['arguments']['node_name'] = req.query['node'];
     if ('version' in req.query)
         data_request['arguments']['version'] = req.query['version'];
 
@@ -124,6 +130,48 @@ router.get('/summary/os', cache(), function(req, res) {
         data_request['arguments']['sort'] = filter.sort_param_to_json(req.query.sort);
     if ('search' in req.query)
         data_request['arguments']['search'] = filter.search_param_to_json(req.query.search);
+
+    execute.exec(python_bin, [wazuh_control], data_request, function (data) { res_h.send(req, res, data); });
+})
+
+/**
+ * @api {get} /agents/no_group Get agents without group
+ * @apiName GetAgentsWithoutGroup
+ * @apiGroup NoGroups
+ *
+ * @apiParam {Number} [offset] First element to return in the collection.
+ * @apiParam {Number} [limit=500] Maximum number of elements to return.
+ * @apiParam {String} [sort] Sorts the collection by a field or fields (separated by comma). Use +/- at the beginning to list in ascending or descending order.
+ * @apiParam {String} [search] Looks for elements with the specified string.
+ * @apiParam {String} [select] List of selected fields.
+ *
+ * @apiDescription Returns a list with the available agents without group.
+ *
+ * @apiExample {curl} Example usage:
+ *     curl -u foo:bar -k -X GET "https://127.0.0.1:55000/agents/no_group?pretty"
+ *
+ */
+router.get('/no_group', cache(), function (req, res) {
+    logger.debug(req.connection.remoteAddress + " GET /agents");
+
+    req.apicacheGroup = "agents";
+
+    var data_request = { 'function': '/agents/no_group', 'arguments': {} };
+    var filters = { 'offset': 'numbers', 'limit': 'numbers', 'sort': 'sort_param', 'select': 'select_param', 'search': 'search_param' };
+
+    if (!filter.check(req.query, filters, req, res))  // Filter with error
+        return;
+
+    if ('offset' in req.query)
+        data_request['arguments']['offset'] = req.query.offset;
+    if ('limit' in req.query)
+        data_request['arguments']['limit'] = req.query.limit;
+    if ('sort' in req.query)
+        data_request['arguments']['sort'] = filter.sort_param_to_json(req.query.sort);
+    if ('search' in req.query)
+        data_request['arguments']['search'] = filter.search_param_to_json(req.query.search);
+    if ('select' in req.query)
+        data_request['arguments']['select'] = filter.select_param_to_json(req.query.select);
 
     execute.exec(python_bin, [wazuh_control], data_request, function (data) { res_h.send(req, res, data); });
 })