+++ CHANGELOG.md
@@ -4,6 +4,36 @@ All notable changes to this project will be documented in this file.
 ## [v3.3.0]
 ### Added
 - Improved agent registration/removal bash script ([#71](https://github.com/wazuh/wazuh-api/pull/71)).
+## [v3.2.3]
+### Added
+- New API requests:
+    * `GET/rules/gdpr` ([#78](https://github.com/wazuh/wazuh-api/pull/78)).
+    * `GET/agents/no_group`.
+    * `GET/cluster/healthcheck`.
+    * `GET/cluster/nodes/:node_name`.
+- A parameter in request `GET/rules` to filter by GDPR requirements ([#78](https://github.com/wazuh/wazuh-api/pull/78)).
+- Parameters in `GET/cluster/nodes`: `search`, `sort`, `offset`, `limit`, `select`. And a new filter: `type`.
+- A parameter in request `GET/agents` to filter agents by cluster nodes.
+### Changed
+- Output of `GET/nodes`: Added a new attribute `version`.
+
+### Removed
+- The following requests have been removed: 
+    - `GET/cluster/agents`: Duplicated request (`GET/agents`).
+    - `GET/cluster/node`: Duplicated request (`GET/config`).
+    - `GET/cluster/files`: It will not be available in this version of the cluster.
+
+## [v3.2.2]
+### Added
+- Added an option in `config.js` to run the API with root privileges for debug purposes and troubleshooting. The API runs as ossec by default. ([#68](https://github.com/wazuh/wazuh-api/pull/68))
+### Changed
+- Changed mode from 750 to 660 in `/configuration/auth/user` file after installing it.
+
+
+## [v3.2.1]
+
+There are no changes for Wazuh API in this version.
+
 
 ## [v3.2.0]
 ### Added